+++ songbird/src/component/option/checkbox/styles.scss
